added wrapper algo, wip test
